requirements.txt
crewai
python-dotenv
requests
litellm
google-search-results